huwei_0806 部署微服务接口 [后台服务体系，应用Api]

+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/domain/Task.java
@@ -453,7 +453,7 @@ public class Task extends EntityMP implements Serializable {
     @JSONField(name = "statusorder")
     @JsonProperty("statusorder")
     @ApiModelProperty("排序")
-    private Integer statusorder;
+    private Long statusorder;
     /**
      * 联系人
      */
